remove unused functions, consolidate inference code in one file
