+++ model/inference.py
@@ -2,15 +2,30 @@ import torch
 from jaxtyping import Float, Int
 from transformers import AutoModelForCausalLM, BatchEncoding
 from transformers.tokenization_utils_fast import PreTrainedTokenizerFast
+from typing import Tuple
 
 from constants import DEVICE
-from david.sampler import generate_with_logprobs, generate_with_logprobs_2
 from model.args import AZRArgs
 
 
 
 def generate_without_grads(model: torch.nn.Module, inputs: BatchEncoding, tokenizer: PreTrainedTokenizerFast, max_new_tokens: int, device: torch.device) -> \
         tuple[Int[torch.Tensor, "batch_size max_response_len"], Float[torch.Tensor, "batch_size max_response_len"]]:
+    """
+    Generate text completions without computing gradients.
+    
+    Args:
+        model: The language model to use for generation
+        inputs: Tokenized input batch with input_ids and attention_mask
+        tokenizer: Tokenizer for the model
+        max_new_tokens: Maximum number of new tokens to generate
+        device: Device to run generation on
+        
+    Returns:
+        Tuple containing:
+            - generated_ids: Shape (batch_size, max_response_len) - Generated token IDs (excluding input tokens)
+            - logprobs_per_token: Shape (batch_size, max_response_len) - Log probabilities for each generated token
+    """
 
 
         outputs = model.generate(
@@ -26,46 +41,18 @@ def generate_without_grads(model: torch.nn.Module, inputs: BatchEncoding, tokeni
         generated_ids = outputs.sequences[:, inputs.input_ids.shape[1] :]
 
 
-        # ------------------------------------------------------------------
-        # OLD: scores = torch.cat(outputs.scores, dim=0)     # wrong shape
-        # NEW: stack on a separate time axis (batch, seq_len, vocab)
-        scores = torch.stack(outputs.scores, dim=1)           # (B, T, V)
-
-        logprobs = torch.log_softmax(scores, dim=-1)          # same shape
+        scores = torch.stack(outputs.scores, dim=1)
+        logprobs = torch.log_softmax(scores, dim=-1)
         
-
-        # Gather the log-prob of each generated token
         logprobs_per_token = logprobs.gather(
             dim=-1,
             index=generated_ids.unsqueeze(-1)
-        ).squeeze(-1)                                         # (B, T)
-
-        # ------------------------------------------------------------------
+        ).squeeze(-1)
 
         return generated_ids, logprobs_per_token
 
 
-# returns the str response and the logprobs for the response
-def generate_response(
-        args: AZRArgs,
-        model: AutoModelForCausalLM,
-        tokenizer: PreTrainedTokenizerFast,
-        prompt: str,
-) -> tuple[
-    str,
-    Float[torch.Tensor, "max_response_len"],
-    Int[torch.Tensor, "max_response_len"],
-    Int[torch.Tensor, "prompt_len"],
-    Int[torch.Tensor, "max_response_len"],
-]:
-    responses, logprobs, gen_ids, prompt_ids, attention_masks = generate_response_bulk(
-        args, model, tokenizer, [prompt]
-    )
-    return responses[0], logprobs[0], gen_ids[0], prompt_ids[0], attention_masks[0]
 
-
-# for each prompt in the list, returns a tuple of lists: (list of str responses, list of logprobs tensors)
-# returns responses (list of strings), logprobs (of the gen responses), generated_ids (tokens of the generated response), inputs.input_ids (tokenized input prompts)
 @torch.no_grad()
 def generate_response_bulk(
         args: AZRArgs,
@@ -79,6 +66,23 @@ def generate_response_bulk(
     Int[torch.Tensor, "batch_size prompt_len"],
     Int[torch.Tensor, "batch_size max_response_len"],
 ]:
+    """
+    Generate text completions for a batch of prompts without computing gradients.
+    
+    Args:
+        args: Configuration arguments containing max_prompt_length and max_response_length
+        model: The language model to use for generation
+        tokenizer: Tokenizer for the model
+        prompts: List of input prompt strings
+        
+    Returns:
+        Tuple containing:
+            - responses: List of decoded response strings
+            - logprobs: Shape (batch_size, max_response_len) - Log probabilities for generated tokens
+            - generated_ids: Shape (batch_size, max_response_len) - Generated token IDs (padded/truncated)
+            - input_ids: Shape (batch_size, prompt_len) - Tokenized input prompts
+            - attention_masks: Shape (batch_size, max_response_len) - Masks for valid (non-padded) positions
+    """
 
 
 
@@ -151,39 +155,100 @@ def generate_response_bulk(
     return responses, logprobs, generated_ids, inputs.input_ids, attention_masks
 
 
-def generate_response_bulk_with_grads(
-        args: AZRArgs,
-        model: AutoModelForCausalLM,
-        tokenizer: PreTrainedTokenizerFast,
-        prompts: list[str],
-) -> tuple[
-    Float[torch.Tensor, "batch_size max_response_len d_vocab"],
-    Float[torch.Tensor, "batch_size max_response_len"],
-    Int[torch.Tensor, "batch_size max_response_len"],
-    Int[torch.Tensor, "batch_size max_response_len d_vocab"],
+def generate_with_logprobs(model: AutoModelForCausalLM,
+                           tokenizer: PreTrainedTokenizerFast,
+                           prompts: list[str],
+                           args: AZRArgs,
+                           debug: bool = False
+                           ) -> Tuple[
+    Int[torch.Tensor, "batch max_new_tokens"],
+    Int[torch.Tensor, "batch max_new_tokens"],
+    Float[torch.Tensor, "batch max_new_tokens"]
 ]:
-    completion_ids, all_logprobs, logprobs_per_token, attention_masks = generate_with_logprobs_2(
-        model, tokenizer, prompts, args
-    )
+    """
+    Generate text completions for a batch of prompts and compute log probabilities.
 
-    return all_logprobs, logprobs_per_token, completion_ids, attention_masks
+    Args:
+        model: The language model to use for generation
+        tokenizer: The tokenizer to use for encoding/decoding
+        prompts: List of input prompt strings
+        args: AZRArgs object containing generation parameters (max_response_length, rollout_top_p, rollout_temperature)
+        debug: Whether to print debug information
+
+    Returns:
+        Tuple containing:
+            - completion_ids: Shape (batch, max_response_length) - Generated token IDs (padded/truncated)
+            - attention_mask: Shape (batch, max_response_length) - Attention mask up to EOS token
+            - logprobs_per_token: Shape (batch, max_response_length) - Log probabilities for generated tokens
+    """
+
+    inputs = tokenizer(prompts,
+                       return_tensors="pt",
+                       padding=True,
+                       padding_side="left")
+
+    prompt_len = inputs.input_ids.shape[1]
+
+    inputs = inputs.to(DEVICE)
+
+    input_ids = model.generate(**inputs,
+                               max_new_tokens=args.max_response_length,
+                               use_cache=True,
+                               do_sample=True,
+                               top_p=args.rollout_top_p,
+                               temperature=args.rollout_temperature)
+
+    logits = model(input_ids).logits
+
+    if debug:
+        print(f"{logits.shape=}")
+        print(f"{prompt_len=}")
+
+    logits = logits[:, prompt_len - 1:-1]
+
+    if debug:
+        print(f"{logits.shape=}")
+
+    logprobs = torch.log_softmax(logits, dim=-1)
+    completion_ids = input_ids[:, prompt_len:]
+
+    logprobs_per_token = torch.gather(logprobs, dim=-1, index=completion_ids.unsqueeze(-1)).squeeze(-1)
+
+    eos_mask = (completion_ids == tokenizer.eos_token_id)
+    eos_positions = torch.argmax(eos_mask.int(), dim=-1)
+    # Handle cases where no EOS token is found (argmax returns 0 even if no match)
+    no_eos_mask = ~eos_mask.any(dim=-1)
+    eos_positions[no_eos_mask] = completion_ids.shape[1]  # Use sequence length if no EOS found
+
+    # Generate attention mask that pays attention to everything up to EOS token (vectorized)
+    batch_size, response_len = completion_ids.shape
+    position_indices = torch.arange(response_len, device=completion_ids.device).unsqueeze(0).expand(batch_size, -1)
+    attention_mask = (position_indices <= eos_positions.unsqueeze(1)).int()
+
+    # Pad/truncate to max_response_length for consistent tensor shapes
+    target_length = args.max_response_length
+    current_length = logprobs_per_token.size(1)
+
+    if debug:
+        print(f"before : {logprobs_per_token.shape=}")
+        print(f"before : {attention_mask.shape=}")
+
+    if current_length < target_length:
+        # Pad with zeros
+        pad_size = target_length - current_length
+        logprobs_per_token = torch.nn.functional.pad(logprobs_per_token, (0, pad_size), value=0)
+        attention_mask = torch.nn.functional.pad(attention_mask, (0, pad_size), value=0)
+        completion_ids = torch.nn.functional.pad(completion_ids, (0, pad_size), value=0)
+
+    elif current_length > target_length:
+        raise ValueError(f"Generated sequence is longer than the target length: {current_length} > {target_length}")
+
+    assert logprobs_per_token.shape == attention_mask.shape == completion_ids.shape, f"inference.py: {logprobs_per_token.shape=} {attention_mask.shape=} {completion_ids.shape=}"
+
+    if debug:
+        print(f"{logprobs_per_token.shape=}")
+        print(f"{attention_mask.shape=}")
+
+    return completion_ids, attention_mask, logprobs_per_token
 
 
-# def remove_dvocab_from_logprobs(
-#         logprobs: Float[torch.Tensor, "role task batch_size max_response_len vocab_size"],
-#         tokens: Int[torch.Tensor, "role task batch_size max_response_len"],
-# ) -> Float[torch.Tensor, "role task batch_size max_response_len"]:
-#     """
-#     Extract log probabilities for generated tokens.
-#
-#     Args:
-#         logprobs: Log probabilities for all tokens in the vocabulary
-#         tokens: Generated token IDs (already excluding prompt)
-#
-#     Returns:
-#         Log probabilities for each generated token
-#     """
-#
-#     # Extract log probabilities for the actual generated tokens
-#     # Use gather to select the log prob for each generated token
-#     return torch.gather(logprobs, dim=-1, index=tokens.unsqueeze(-1)).squeeze(-1)
